requirements.txt
docstrange
python-dotenv
pypdf
requests
pytest
